introduce types for usage config
