+++ modules/usage-s3/module_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/weaviate/weaviate/usecases/config"
 	"github.com/weaviate/weaviate/usecases/config/runtime"
 	common "github.com/weaviate/weaviate/usecases/modulecomponents/usage"
+	usagetypes "github.com/weaviate/weaviate/usecases/modulecomponents/usage/types"
 )
 
 func TestModule_Name(t *testing.T) {
@@ -47,7 +48,7 @@ func TestModule_Init_Success(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			ScrapeInterval: runtime.NewDynamicValue(5 * time.Minute),
 			PolicyVersion:  runtime.NewDynamicValue("2025-06-01"),
 		},
@@ -75,7 +76,7 @@ func TestModule_Init_MissingBucket(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			ScrapeInterval: runtime.NewDynamicValue(5 * time.Minute),
 			PolicyVersion:  runtime.NewDynamicValue("2025-06-01"),
 		},
@@ -95,7 +96,7 @@ func TestModule_Init_MissingEnvVars(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			ScrapeInterval: runtime.NewDynamicValue(5 * time.Minute),
 			PolicyVersion:  runtime.NewDynamicValue("2025-06-01"),
 		},
@@ -120,7 +121,7 @@ func TestModule_Init_MissingHostname(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "", // Empty hostname
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			ScrapeInterval: runtime.NewDynamicValue(5 * time.Minute),
 			PolicyVersion:  runtime.NewDynamicValue("2025-06-01"),
 		},
@@ -146,7 +147,7 @@ func TestModule_Init_ConfigurationParsing(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			ScrapeInterval: runtime.NewDynamicValue(5 * time.Minute), // existing config takes priority
 			PolicyVersion:  runtime.NewDynamicValue("2025-01-01"),    // existing config takes priority
 		},
@@ -183,7 +184,7 @@ func TestModule_Init_InvalidScrapeInterval(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			PolicyVersion: runtime.NewDynamicValue("2025-06-01"),
 		},
 	}
@@ -204,7 +205,7 @@ func TestModule_BuildS3Config(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			S3Bucket:      runtime.NewDynamicValue("test-bucket"),
 			S3Prefix:      runtime.NewDynamicValue("test-prefix"),
 			PolicyVersion: runtime.NewDynamicValue("2025-06-01"),
@@ -226,7 +227,7 @@ func TestModule_BuildS3Config_EmptyValues(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			// All values are nil/empty
 		},
 	}
@@ -253,7 +254,7 @@ func TestModule_SetUsageService(t *testing.T) {
 		Cluster: cluster.Config{
 			Hostname: "test-node",
 		},
-		Usage: config.UsageConfig{
+		Usage: usagetypes.UsageConfig{
 			ScrapeInterval: runtime.NewDynamicValue(5 * time.Minute),
 			PolicyVersion:  runtime.NewDynamicValue("2025-06-01"),
 		},
@@ -357,7 +358,7 @@ func TestParseS3Config(t *testing.T) {
 			}
 
 			config := &config.Config{
-				Usage: config.UsageConfig{
+				Usage: usagetypes.UsageConfig{
 					ScrapeInterval: runtime.NewDynamicValue(5 * time.Minute),
 					PolicyVersion:  runtime.NewDynamicValue("2025-06-01"),
 				},
@@ -464,7 +465,7 @@ func TestParseCommonUsageConfig(t *testing.T) {
 			}
 
 			config := &config.Config{
-				Usage: config.UsageConfig{},
+				Usage: usagetypes.UsageConfig{},
 			}
 
 			// Set existing values if specified
